We can use higher order component as long as we do it in a server component
